Title: 입력받은 수의 합과 평균, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240608/입력받은 수의 합과 평균/sum-and-average-of-the-inputs.cpp
@@ -11,6 +11,6 @@ int main() {
 	}
 	cout << fixed;
 	cout.precision(1);
-	cout << s << " " << s / a;
+	cout << (int)s << " " << s / a;
 	return 0;
 }
